spicy: show the correct ungrab key combination
The gtk accelerator for ungrab is useless, since it has to be handled
by the spice widget only. It could be useful to still show the ungrab
key sequence in the menu (for help), but unfortunately, spice-gtk grab
sequence syntax is not the same as gtk accelerator syntax, and that
would be needlessly complicated to handle.

Also correctly show the configured sequence in the status bar when the
widget has the grab.

https://bugzilla.redhat.com/show_bug.cgi?id=851090