win-usb-driver: use usbclerk new message: USB_CLERK_DRIVER_SESSION_INSTALL
With this message usbclerk keeps a list of devices for which
a libusb driver was installed (per connection).
When a spice-gtk client exits, the connection is closed, and
usbclerk uninstalls the driver for all devices in the list.

That means we need to keep the connection open, so added
the win-usb driver installer to usb-device-manager's priv.

This prevents the case were the user exits the client, while a usb
device is connected to the guest, and can not use the device from
the client machine.

rhbz#869542