channel-inputs: Fix sending 00 scancodes to guests with scancode cap
The code for handling single key up / down events in spice-server is:
        SpiceMsgcKeyDown *key_down = (SpiceMsgcKeyDown *)buf;
        uint8_t *now = (uint8_t *)&key_down->code;
        uint8_t *end = now + sizeof(key_down->code);
        for (; now < end && *now; now++) {
            kbd_push_scan(keyboard, *now);
        }

Notice the *now, which makes sure that no scancodes with the value 0 get
send! But the new SPICE_MSGC_INPUTS_KEY_SCANCODE in the server does:

        uint8_t *code = (uint8_t *)buf;
        for (i = 0; i < size; i++) {
            kbd_push_scan(keyboard, code[i]);
        }

And thus will push any 0 bytes in the buffer. Resulting in these message
in the guest:

atkbd serio0: Unknown key pressed (translated set 2, code 0x0 on isa0060/serio0).
atkbd serio0: Use 'setkeycodes 00 <keycode>' to make it known.

Rather then making the server skip 0 bytes I believe it is better to just
make spice-gtk not send these in the first place, which is what this patch
does.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>