widget: don't redraw server mouse until moved
When switching between client mode and server mode, the pointer is
being invalidated on each display and the cursor will end up being
drawn on both. Since there is no information on which display the
cursor is supposed to be until a move is received, hide the cursor
until it actually moves.