usb-device-manager: Fix coldplug race
It is possible for us to see a device show up twice, if it gets plugged
in between us starting listening for new devices and doing "coldplug", then
it will get added once from the coldplug code, and then again from from
the hotplug code path. We already have code checking for this, but the check
is only compiled in under Windows -> Remove the #ifdef to also catch this
under Linux.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>