spice-pulse: adjust the playback latency when the min-latency property changes
If the current latency is smaller than the new min-latency
value, we cork the playback till the target latency is achieved.

Note: I didn't modify the prebuf configuration and used
pa_stream_prebuf, because pulse updated the prebuf only if
I set both prebuf and tlength to be target_latency*2. Then,
due to the tlength growth, each time the playback latency deviated
from the target latency, an underflow occurred. Since the latency
that is computed by the server is not exact and is based on its
current evaluation of the bit-rate, which is dynamic, it is better not
to change the tlength (in order to avoid unnecessary underflows).