spice-widget: Reconfiguring the grab is only needed on win32
Commit 8a7e72e3 "widget: regrab when widget is reconfigured" adds an
ungrab + grab call to the configure event handling code. Because:
"On Windows, we need to update the cursor clip. Call ungrab&grab to update it."

But on X11 this is not needed, see man XGrabPointer, which explains that
the grab automatically adjusts to window resizing.

Not only is it not needed it is also racy, causing spice-gtk based
apps to log messages like:
(remote-viewer:9935): GSpice-WARNING **: pointer grab failed 3

This patch fixes this by disabling the ungrab + re-grab on non-win32.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>