usb-device-manager: Windows: identify devices by vid:pid instead of bus.address
rhbz#842816

Sometimes bus.address of a USB device changes upon WinUSB driver installation
for that device. This makes bus.address not a good identifier to use when
running on Windows machines.

Instead this patch makes usb-device-manager, when running on a Windows client,
identify devices by their vid:pid.

What changes were made in this patch (in addition to previous patches):
- vid:pid are asked from the udev.
- match functions that compare vid:pid were added
- when comparing devices vid:pid are used.

This also means that a scenario where two USB devices with the same vid:pid
on a Windows client is not supported. However there was a problem with this
scenario before as on Windows drivers for (specific) USB devices are
installed based on their vid:pid.