channel-main: Don't g_warn when not finding a file-xfer-id
Since the agent channel can have a significant latency, it is possible
for a cancel send from the client to the agent and a status message from
the agent to cross each other. If this happens then the file-xfer will no
longer be on the list.

Printing a g_warning on this (rare but triggerable) condition will only get
us hard to debug bug reports, so turn it into a SPICE_DEBUG.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>