Release pointer grab on grab-broken
On windows, the client receives a WM_KILLFOCUS event which generates
solely a keyboard grab-broken event.

This event is received when pressing ctrl-alt-del (to show up the task
manager), and we need to release the pointer grab and clip region in
this case for the client to be usable.

This also clear the clipping region when the client pops-up a dialog.

Since keyboard focus is a pre-requisite for pointer grab, it sounds
logical to release the pointer grab if losing keyboard focus, but for
now, we just release it in grab-broken, as a result of discussion on
the ML.

Fixes:
https://bugzilla.redhat.com/show_bug.cgi?id=857114
https://bugzilla.redhat.com/show_bug.cgi?id=922818