channel-main: Don't send empty monitor config on main_init with agent
Currently we send a monitor-config on any agent_start, including sending it if
the agent is already started when the main channel connects, but when the main
channel gets initialized, the display channels aren't intialized yet, so our
monitor config will be empty. Resulting in the Linux agent logging:

spice-vdagent[1285]: err: client sent config with all monitors disabled

This patch fixing this by only sending our monitor-config to the agent when
it (re)starts later on.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>