widget: don't grab mouse when switching to server mode
spice-gtk used to try taking the grab when switching to server mode, so
that the current mouse interaction (for example a drag), isn't
interrupted if the agent dies. However, shutting down a VM will
automatically try taking the grab when there is no tablet input, and
agent exits.

Some users are complaining about it (especially when they don't know the
ungrab combination).

Instead, let's try to keep the grab only if we have currently a mouse
button pressed.

https://bugzilla.redhat.com/show_bug.cgi?id=873272