main: do not send monitors config if some are missing
Spice-gtk does a bit of client-side work by optionnally delaying sending
the monitor configuration to the guest automatically. However, the
client may be slow to set all the monitors, so teach the timer to not
fire the event unless at least the number of monitors set explicitely
enabled or disabled matches the number of display channels.

This avoid some configuration races when connecting to a multi-channel
display server which is slow to set up.