win32: don't block key up events
The win32 ll keyboard hook avoid Windows and other application to
receive global keyboard events. But some key combinations can't be
filtered, such as Win+L. However, the windows lock screen doesn't catch
that the Win key is released, when virt-viewer still holds the hook and
filters it.

So pressing Win+L quickly will lock the screen, but some key press in
the password entry will still be handled as if the Win key was pressed,
such as Win+P or Win+U and probably other, and prevents user from typing
his password.

The only working solution I could find is to just let go all the release
key events in the hook. There doesn't seem any drawback with that.

https://bugzilla.redhat.com/show_bug.cgi?id=917986