Only use GOBJECT_INTROSPECTION_CHECK when available
Current el6 systems don't have gobject-introspection packages,
so the GOBJECT_INTROSPECTION_CHECK m4 macro is not available
on such systems. This makes it difficult to run autogen.sh
on these systems.
This commit uses m4_ifdef to check if this macro is available
before trying to use it, and makes sure introspection is
disabled when GOBJECT_INTROSPECTION_CHECK could not be found.
This allows to compile spice-gtk from git on el6 systems
with:
./autogen.sh --with-gtk=2.0 --disable-gtk-doc --disable-vala
             --disable-controller --disable-werror