build-sys: Add real autodetection of python
configure help text says that by default python will be used automatically
if appropriate, however this only means that on !windows and !gtk3, we will
force build of the python bindings, and fail if any needed package to build
them is missing.

This commit improves on the current checks so that when autodetecting
python, if a package is missing then build of the bindings is silently
disabled rather than failing.