pulse: do not abort on corking no stream
There is no guarantee that a stream actually exist when min-latency
change is notified.

Check that there is an actual stream before calling cork().  All callers
where previously checking that stream existed. Add a pre-condition to
verify argument and spot that error is from spice-gtk itself.

Fixes the following crash:
 #1  0x0000003c12e34105 in abort () at abort.c:92
 #2  0x0000003c2c223180 in pa_stream_is_corked (s=0x0) at
 pulse/stream.c:2536
 #3  0x0000003c2c648cb7 in stream_cork (pulse=<value optimized out>,
 s=0x7fbb38, with_flush=0) at spice-pulse.c:227
 #4  0x0000003c2c649989 in playback_min_latency_changed (object=<value
 optimized out>, pspec=<value optimized out>, data=0x7fbad0) at
 spice-pulse.c:674

https://bugzilla.redhat.com/show_bug.cgi?id=1032785