clipboard: prevent reentering main loop if main channel is busy
The main channel must be idle to avoid entering the nested loop,
or it will never reach the condition to leave it.

This should not happen. It can happen currently when the remote sends a
clipboard request while the clipboard is grabbed by the remote. In this
case, while the clipboard-request signal is emitted by main channel, the
clipboard_get() loop is entered, but the main coroutine will not be
woken up to proceed with the request, the main channel will remain
"frozen" and it won't be possible to leave cleanly from the inner loop.
The application appears to be frozen, because it can't quit properly.

Related to:
https://bugzilla.redhat.com/show_bug.cgi?id=1083489