Revert "clipboard: prevent reentering main loop if main channel is busy"
This change broke primary clipboard copying from remote to client, on win32.

There is a pending ::clipboard-grab emission, so the channel is not
idle, and it will handle emission and resume in the inner loop without
issue. We should really handle this situation with care though, and the
best would be to teach gtk+ to do async clipboard request.

This reverts commit e3efa8cec51045d87a14a7e4f921c24ad5caffb1.