Prefer using g_malloc0()/g_free()
As we already depend on GLib, let's use g_{malloc,new}0() instead of the
standard malloc() or the spice_{malloc,new}*() and g_free() instead of
the standard free() when possible.
Memory allocated by other libraries using malloc() should still be freed
by free().
As a side effect of the changes, we are muting a few warnings caught by
coverity.