Ensure '\0' is not part of text clipboard data
On Windows, with some versions of gtk+, GtkSelectionData::length
will include the final '\0'.
When a string with this trailing '\0' is pasted in some linux
applications, it will be pasted as <NIL> or as an invisible character,
which is unwanted.

This commit ensures the length we send to the agent does not
include any trailing '\0'.

https://bugzilla.redhat.com/show_bug.cgi?id=1090122