win32: fix coroutine assertion when switching to TLS port
Since v0.21-35-gbaa51c5, there are some new coroutine preconditions.

The winfiber implementation can reach, when the structure isn't cleared,
which happens when switching channel connection to TLS port and the
coroutine is recycled.

CRITICAL **: coroutine_yieldto: assertion '!to->exited' failed

In a near future, it would be nice to use the safer implementation
proposed for glib, see bgo#719362 (using a private copy)