audio: use weak references to channel
The audio channels are currently referenced and released on channel
close events. However, this event may not happen if the channel never
was connected. Keeping channels alive also prevent session from
finishing.

By not holding the ref, the channel can go to dispose
when it is no longer needed, and the session can be disposed too.