Release keyboard grab using keyboard shortcut
This commit adds the ability to release the keyboard grab when
the release keys (ctrl+alt) are pressed and released. It allows
to use keyboard shortcuts (eg alt+tab, alt+f4) on the client.

The keyboard is grabbed again when the release keys are pressed
and released or when the mouse moves.

https://bugs.freedesktop.org/show_bug.cgi?id=85331