Fix empty clipboard check
SpiceGtkSession::clipboard_received_cb starts by checking if the
length of the X selection data is not 0. However, right after this check,
if gtk_selection_data_get_length() returned -1, it decides it got an
empty clipboard, sets the selection length to 0, and does not return
early.
This commit reworks the len == 0 / len == -1 checks to make sure we
always return early when we get no data from the clipboard.